requirements.txt
aiohttp==3.7.4.post0
click==7.1.2
dacite==1.6.0
gql @ git+https://github.com/graphql-python/gql@af8f22362d70a6989a703bca5c941b06065d69c2
python-dotenv==0.19.1
spotipy==2.19.0
Pygments==2.10.0
retry==0.9.2
